Rework commande info, needs clean up

+++ CircuitsCourtsApi/src/main/java/com/lifat/CircuitsCourtsApi/model/CommandeDetail.java
@@ -28,6 +28,6 @@ public class CommandeDetail {
 
     @Min(0)
     @Column(name = "quantite")
-    private Double quantite;
+    private Float quantite;
 
 }
